pyproject.toml
[tool.poetry]
name = "backtrader-2-modular"
version = "0.1.0"
description = "Modular backtesting system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5.0"
numpy = "^1.23.0"
pyyaml = "^6.0"
matplotlib = "^3.5.0"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^22.3.0"
flake8 = "^4.0.0"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fetch-data = "src.data.fetch_data:main"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
